Cancel timers during unit testing
