+++ tester/tester_test.go
@@ -148,90 +148,6 @@ func activityLongRunning(ctx context.Context) (int, error) {
 	return 42, nil
 }
 
-func Test_Timer(t *testing.T) {
-	tester := NewWorkflowTester[timerResult](workflowTimer)
-	start := tester.Now()
-
-	tester.Execute()
-
-	require.True(t, tester.WorkflowFinished())
-	wr, _ := tester.WorkflowResult()
-	require.True(t, start.Equal(wr.T1))
-
-	e := start.Add(30 * time.Second)
-	require.True(t, e.Equal(wr.T2), "expected %v, got %v", e, wr.T2)
-}
-
-type timerResult struct {
-	T1 time.Time
-	T2 time.Time
-}
-
-func workflowTimer(ctx workflow.Context) (timerResult, error) {
-	t1 := workflow.Now(ctx)
-
-	workflow.ScheduleTimer(ctx, 30*time.Second).Get(ctx)
-
-	t2 := workflow.Now(ctx)
-
-	workflow.ScheduleTimer(ctx, 30*time.Second).Get(ctx)
-
-	return timerResult{
-		T1: t1,
-		T2: t2,
-	}, nil
-}
-
-func Test_TimerCancellation(t *testing.T) {
-	tester := NewWorkflowTester[time.Time](workflowTimerCancellation)
-	start := tester.Now()
-
-	tester.Execute()
-
-	require.True(t, tester.WorkflowFinished())
-
-	wfR, _ := tester.WorkflowResult()
-	require.True(t, start.Equal(wfR), "expected %v, got %v", start, wfR)
-}
-
-func workflowTimerCancellation(ctx workflow.Context) (time.Time, error) {
-	tctx, cancel := workflow.WithCancel(ctx)
-	t := workflow.ScheduleTimer(tctx, 30*time.Second)
-	cancel()
-
-	_, _ = t.Get(ctx)
-
-	return workflow.Now(ctx), nil
-}
-
-func Test_TimerRespondingWithoutNewEvents(t *testing.T) {
-	tester := NewWorkflowTester[time.Time](workflowTimerRespondingWithoutNewEvents)
-
-	tester.ScheduleCallback(time.Duration(2*time.Second), func() {
-		tester.SignalWorkflow("signal", "s42")
-	})
-
-	tester.Execute()
-
-	require.True(t, tester.WorkflowFinished())
-
-	_, err := tester.WorkflowResult()
-	require.Empty(t, err)
-}
-
-func workflowTimerRespondingWithoutNewEvents(ctx workflow.Context) error {
-	workflow.ScheduleTimer(ctx, 1*time.Second).Get(ctx)
-
-	workflow.Select(
-		ctx,
-		workflow.Receive(workflow.NewSignalChannel[any](ctx, "signal"), func(ctx workflow.Context, signal any, ok bool) {
-			// do nothing
-		}),
-	)
-
-	return nil
-}
-
 func Test_Signals(t *testing.T) {
 	tester := NewWorkflowTester[string](workflowSignal)
 	tester.ScheduleCallback(time.Duration(5*time.Second), func() {
