+++ tester/tester_timers_test.go
@@ -1,0 +1,135 @@
+package tester
+
+import (
+	"testing"
+	"time"
+
+	"github.com/cschleiden/go-workflows/workflow"
+	"github.com/stretchr/testify/require"
+)
+
+func Test_Timer(t *testing.T) {
+	tester := NewWorkflowTester[timerResult](workflowTimer)
+	start := tester.Now()
+
+	tester.Execute()
+
+	require.True(t, tester.WorkflowFinished())
+	wr, _ := tester.WorkflowResult()
+	require.True(t, start.Equal(wr.T1))
+
+	e := start.Add(30 * time.Second)
+	require.True(t, e.Equal(wr.T2), "expected %v, got %v", e, wr.T2)
+}
+
+type timerResult struct {
+	T1 time.Time
+	T2 time.Time
+}
+
+func workflowTimer(ctx workflow.Context) (timerResult, error) {
+	t1 := workflow.Now(ctx)
+
+	workflow.ScheduleTimer(ctx, 30*time.Second).Get(ctx)
+
+	t2 := workflow.Now(ctx)
+
+	workflow.ScheduleTimer(ctx, 30*time.Second).Get(ctx)
+
+	return timerResult{
+		T1: t1,
+		T2: t2,
+	}, nil
+}
+
+func Test_TimerCancellation(t *testing.T) {
+	tester := NewWorkflowTester[time.Time](workflowTimerCancellation)
+	start := tester.Now()
+
+	tester.Execute()
+
+	require.True(t, tester.WorkflowFinished())
+
+	wfR, _ := tester.WorkflowResult()
+	require.True(t, start.Equal(wfR), "expected %v, got %v", start, wfR)
+}
+
+func workflowTimerCancellation(ctx workflow.Context) (time.Time, error) {
+	tctx, cancel := workflow.WithCancel(ctx)
+	t := workflow.ScheduleTimer(tctx, 30*time.Second)
+	cancel()
+
+	_, _ = t.Get(ctx)
+
+	return workflow.Now(ctx), nil
+}
+
+func Test_TimerSubworkflowCancellation(t *testing.T) {
+	tester := NewWorkflowTester[time.Time](workflowSubWorkflowTimerCancellation)
+	tester.Registry().RegisterWorkflow(timerCancellationSubWorkflow)
+
+	tester.Execute()
+
+	require.True(t, tester.WorkflowFinished())
+
+	_, wfErr := tester.WorkflowResult()
+	require.Empty(t, wfErr)
+}
+
+func workflowSubWorkflowTimerCancellation(ctx workflow.Context) error {
+	_, err := workflow.CreateSubWorkflowInstance[any](ctx, workflow.DefaultSubWorkflowOptions, timerCancellationSubWorkflow).Get(ctx)
+
+	// Wait long enough for the second timer in timerCancellationSubWorkflow to fire
+	workflow.Sleep(ctx, 20*time.Second)
+
+	return err
+}
+
+func timerCancellationSubWorkflow(ctx workflow.Context) error {
+	tctx, cancel := workflow.WithCancel(ctx)
+
+	t := workflow.ScheduleTimer(ctx, 2*time.Second)
+	t2 := workflow.ScheduleTimer(tctx, 10*time.Second)
+
+	workflow.Select(
+		ctx,
+		workflow.Await(t, func(ctx workflow.Context, f workflow.Future[struct{}]) {
+			// Cancel t2
+			cancel()
+		}),
+		workflow.Await(t2, func(ctx workflow.Context, f workflow.Future[struct{}]) {
+			// do nothing here, should never fire
+			panic("timer should have been cancelled")
+		}),
+	)
+
+	return nil
+}
+
+func Test_TimerRespondingWithoutNewEvents(t *testing.T) {
+	tester := NewWorkflowTester[time.Time](workflowTimerRespondingWithoutNewEvents)
+
+	tester.ScheduleCallback(time.Duration(2*time.Second), func() {
+		tester.SignalWorkflow("signal", "s42")
+	})
+
+	tester.Execute()
+
+	require.True(t, tester.WorkflowFinished())
+
+	_, err := tester.WorkflowResult()
+	require.Empty(t, err)
+}
+
+func workflowTimerRespondingWithoutNewEvents(ctx workflow.Context) error {
+	workflow.ScheduleTimer(ctx, 1*time.Second).Get(ctx)
+
+	workflow.Select(
+		ctx,
+		workflow.Receive(workflow.NewSignalChannel[any](ctx, "signal"), func(ctx workflow.Context, signal any, ok bool) {
+			// do nothing
+		}),
+	)
+
+	return nil
+}
